requirements.txt
absl-py==0.9.0
alembic==1.4.1
aniso8601==8.0.0
astor==0.8.1
astunparse==1.6.3
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
cycler==0.10.0
dominate==2.5.1
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
flask-crontab==0.1.2
Flask-Login==0.5.0
Flask-Migrate==2.5.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gast==0.3.3
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.27.2
h5py==2.10.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Mako==1.1.2
Markdown==3.2.1
MarkupSafe==1.1.1
marshmallow==3.5.1
matplotlib==3.2.0
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.2.0
pandas==1.0.1
pathlib2==2.3.5
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.1
PyJWT==1.7.1
PyMySQL==0.9.3
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
pytz==2019.3
PyYAML==5.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
seaborn==0.10.0
six==1.14.0
SQLAlchemy==1.3.15
tb-nightly==2.3.0a20200424
tensorboard==2.1.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.1.0
tensorflow-docs==0.0.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
tf-estimator-nightly==2.3.0.dev2020042401
tf-nightly==2.2.0.dev20200424
urllib3==1.25.8
visitor==0.1.3
webargs==6.0.0
Werkzeug==1.0.0
wrapt==1.12.1
WTForms==2.2.1
git+https://github.com/tensorflow/docs
